+++ src/main/java/gin/melec/SplitDown.java
@@ -26,17 +26,17 @@ import static gin.melec.AbstractSplit.WINDOW;
 public class SplitDown extends AbstractSplit {
 
     @Override
-    protected float distanceTo(final Vertex vertex) {
+    protected double distanceTo(final Vertex vertex) {
         return Math.abs(this.position - vertex.getY());
     }
 
     @Override
-    protected final long xPosition() {
+    protected final double xPosition() {
         return 0;
     }
 
     @Override
-    protected final long yPosition() {
+    protected final double yPosition() {
         return position;
     }
