+++ src/main/java/gin/melec/SplitRight.java
@@ -26,17 +26,17 @@ import static gin.melec.AbstractSplit.WINDOW;
 public class SplitRight extends AbstractSplit {
 
     @Override
-    protected float distanceTo(final Vertex vertex) {
+    protected double distanceTo(final Vertex vertex) {
         return Math.abs(this.position - vertex.getX());
     }
 
     @Override
-    protected final long xPosition() {
+    protected final double xPosition() {
         return position;
     }
 
     @Override
-    protected final long yPosition() {
+    protected final double yPosition() {
         return 0;
     }
